requirements.txt
streamlit==1.19.0
watchdog==2.3.0
pandas==1.3.0
pytest==6.2.4
scikit-learn==1.0.2
plotly-express==0.4.1